requirements.txt
streamlit
pdf2image
pytesseract
Pillow